requirements.txt
GitPython==2.1.5
Jinja2==2.9.6
